requirements.txt
flask
groq
PyPDF2
google-api-python-client
google-auth